ci: refactor github actions
- run golangci-lint instead of staticcheck & go vet only
- build the package during CI

Building and testing is done via multiple go versions, golangci-lint is only run
against the latest go-version that we use.

The tests are still run via the coverage job unnecessarily a second time, as
before.

Because github actions does not support standard YAML achors
(https://github.com/actions/runner/issues/1182) we have to duplicate the
definitions of the go versions.
I'm not aware of a way to prevent those duplicates.