xtypes: fixing issues with default values (#9)
Some XTypes, like xtypes.OneOf, might require a default value to be specified
for mandatory parameters. If for example the list of choices is do, re, mi and
the parameter is marked as optional, then the default value must be one of the
provided values. If the parameter is not optional, then the default value will
not be used, meaning that it doesn't mater what the value is. The logic for
handling this condition existed, but it was not correct. The test provider,
for example, only supported setting a parameter, but not unsetting it.

A second problem solved on this PR is that an optional XType might have a
value, then the value might stop being provided (for example, a file that
provides the value is deleted). In this case, the value must changed back to
the default value. This condition was also covered by code, but it was not
working correctly.

Finally, a log message was being shown every time the value of a parameter
configured using OneOf was being used. It was also calling the E (error)
helper function, but that was logging it as info. The log message caused no
side effects, but was confusing.

This PR fixes the issues above by:

- Validating the default values for all optional parameter provided using
  xtypes. This condition is a coding error, so is handled by panicking
  MustParse. Test coverage is included.
- Test provider now supports unsetting a parameter, and a test for reverting
  to default was included;
- The test logger was changed to cause the test to fail if a log message with
  severity error is logged.

The issues fixed by this PR, except by the log message, affected only dynamic
providers, which are not merged yet.