testing: removing dependencies (#13)
Proteus have only test dependencies, and this is only for tests, to help with
assertions. The kind of task being delegated are quite simple, like asserting
that two values are equal, and the library was designed before go.18, meaning
that many parameters are of type interface{}, resulting in less opportunities
for the compiler to catch errors.

This PR removes the test library and replaces it by an internal test library.
This will remove 23k LOC from the proteus repository.

The test introduces potentially breaking change on the plog package, where
Entry.Caller is changed from pointer to value. Due to go's automatic type
indirection, code that only makes reference to Caller fields will not be
affected, but other things like checking for nil would break.