options: introduce WithValueFormatting option (#15)
This PR introduces an option called WithValueFormatting. It allows specifying
options for pre-processing values before they get used. The initial supported
feature is trimming values.

This is useful for some cases where the value of the parameters contains
undesired spaces, like when a bad template language is used for providing
values, leading to undesired spaces or even trailing new line.