aesthetics: include default values on Dump()
When the value of the parameter are dumped only the provided parameter are
being included in the response. This PR changes it to also include optional
parameters. The optional parameters are identified by including the string
" (default)" at the end of the line.

This PR also makes two small changes:
- `assert` package: more clearly show the difference between two values, when
  the assertion fail
- Renames some example files to have "example_" as a prefix, to make it easier
  to identify which files contain examples and which contains tests.