do not print help output on usage error
When an error happened, like a required parameter was not passed, Proteus was
printing a usage description and a help text.

If an application had a lot of parameter the help output was very long and made
it harder to find the actual error.

In most circumstances the usage description is sufficient to figure out what the
application expects.

- split printing the usage and help output into 2 private functions, help() and
  usage(),
- change parsed.ErrUsage() to only print the usage output,
- move some code that is shared between Usage() and Help() to helper functions,
  to reduce duplication