bugfix: panic if unsupported field type is ignored
A configuration struct may contain some fields that are not intended to be
used for configuration. Those fields can be marked with the name "-", like:

```go
cfg := struct{
    A string
    B func() `param:"-"` // ignore: not a parameter
}{}
```

When this case was used on proteus would check the type of the struct
before checking the name, resulting in a panic about configuration error
on the field B.

This PR fixes the issue by correctly checking if the fields is ignored before
looking into the field type. A test is included to avoid regression.